chore(compiler): split compiler in subpackages
