+++ storyscript/compiler/__init__.py
@@ -1,8 +1,4 @@
 # -*- coding: utf-8 -*-
-from .Compiler import Compiler
-from .Faketree import FakeTree
-from .Lines import Lines
-from .Objects import Objects
-from .Preprocessor import Preprocessor
+from storyscript.compiler.Compiler import Compiler
 
-__all__ = ['Compiler', 'FakeTree', 'Lines', 'Objects', 'Preprocessor']
+__all__ = ['Compiler']
