+++ storyscript/compiler/json/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+from storyscript.compiler.json.JSONCompiler import JSONCompiler
+from storyscript.compiler.json.Lines import Lines
+from storyscript.compiler.json.Objects import Objects
+
+__all__ = ['JSONCompiler', 'Lines', 'Objects']
