+++ storyscript/compiler/semantics/ReturnVisitor.py
@@ -1,15 +1,16 @@
 # -*- coding: utf-8 -*-
 
-from ..exceptions import CompilerError
+from storyscript.exceptions import CompilerError
+
+from .Visitors import SelectiveVisitor
 
 
-class Postprocessor:
+class ReturnVisitor(SelectiveVisitor):
     """
     Checks the return type of functions.
     """
 
-    @classmethod
-    def has_return(cls, tree):
+    def has_return(self, tree):
         if tree.rules and tree.rules.return_statement:
             return True
 
@@ -21,7 +22,7 @@ class Postprocessor:
             # an if requires an else and all statements to have a return
             if tree.if_block.else_block:
                 for block in tree.if_block.children[1:]:
-                    if not cls.has_return(block):
+                    if not self.has_return(block):
                         return False
                 return True
             return False
@@ -29,20 +30,18 @@ class Postprocessor:
         nested_block = tree.nested_block or tree.child(0).nested_block
         if nested_block:
             for block in nested_block.children:
-                if cls.has_return(block):
+                if self.has_return(block):
                     return True
         return False
 
-    @classmethod
-    def function_block(cls, tree):
+    def function_block(self, tree):
         if tree.function_statement.function_output:
-            if not cls.has_return(tree):
+            if not self.has_return(tree):
                 t = tree.function_statement.function_output
                 raise CompilerError('return_required', tree=t)
 
-    @classmethod
-    def process(cls, tree):
-        for block in tree.find_data('block'):
-            if block.function_block is not None:
-                cls.function_block(block.function_block)
-        return tree
+    def block(self, tree):
+        self.visit_children(tree)
+
+    def start(self, tree):
+        self.visit_children(tree)
