+++ storyscript/compiler/semantics/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+from storyscript.compiler.semantics.Semantics import Semantics
+
+
+__all__ = ['Semantics']
